requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-google-genai
langchain-chroma
langchain-community
google-cloud-secret-manager
google-cloud-logging
google-auth
google-cloud-storage
firebase-admin
chroma
pypdf
fpdf
youtube-transcript-api
pytube
python-dotenv
python-pptx
fastapi[standard]
pymupdf
chromadb
pdfplumber
pydantic==1.10.7
langchain_groq
selenium
bs4
firecrawl-py
PyPDF2
python-docx
reportlab
moviepy
assemblyai
